+++ core/data_structures/__init__.py
@@ -1,2 +1,1 @@
 from core.data_structures.graph import Graph, Node, connect_both
-from core.data_structures.transport_structures import Bus, PassengersGroup
