+++ core/simulation/__init__.py
@@ -1,2 +1,5 @@
 from core.simulation import *
+from core.simulation.bus import Bus
+from core.simulation.passenger_group import PassengersGroup
 from core.simulation.simulation import Simulation
+from core.simulation.stop import Stop
