--- core/simulation/bus_collector.py
@@ -1,3 +1,0 @@
-class BusCollector:
-    def __init__(self):
-        pass
