+++ core/simulation/stop.py
@@ -1,0 +1,4 @@
+class Stop:
+    def __init__(self, name):
+        self.name = name
+        self.passengers = []
